Change the data type of the phase to a single precision float. (#68)
With it comes the change to a single precision float in all of tuning. Bench actually changes, so let's test this out first...
ELO   | -0.40 +- 2.74 (95%)
SPRT  | 10.0+0.1s Threads=1 Hash=8MB
LLR   | 2.96 (-2.94, 2.94) [-5.00, 1.00]
Games | N: 34128 W: 9465 L: 9504 D: 15159
http://chess.grantnet.us/test/7088/
BENCH=12241175

+++ core-sdk/src/evaluation/mod.rs
@@ -23,8 +23,8 @@ pub const EG: usize = 1;
 #[derive(Copy, Clone, PartialEq)]
 pub struct EvaluationScore(pub i16, pub i16);
 impl EvaluationScore {
-    pub fn interpolate(self, phase: f64) -> i16 {
-        ((f64::from(self.0) * phase + f64::from(self.1) * (128.0 - phase)) / 128.0) as i16
+    pub fn interpolate(self, phase: f32) -> i16 {
+        ((f32::from(self.0) * phase + f32::from(self.1) * (128.0 - phase)) / 128.0) as i16
     }
 }
 impl Default for EvaluationScore {
@@ -356,7 +356,7 @@ pub fn is_guaranteed_draw(g: &GameState) -> bool {
     }
     false
 }
-pub fn endgame_rescaling(g: &GameState, res: &mut EvaluationScore, phase: f64) {
+pub fn endgame_rescaling(g: &GameState, res: &mut EvaluationScore, phase: f32) {
     let score = res.interpolate(phase);
     let side_ahead = if score >= 0 { WHITE } else { BLACK };
     let side_losing = 1 - side_ahead;
